virtio-blk: fix hardware used index size
According to the virtio spec used index must be uint16.
As the result of this bug queues created with non zero used
index were not working correctly.

(cherry picked from commit 332a2b4e438997d2fc4be0916ce619025d3e4b35)

+++ src/snap_virtio_blk.h
@@ -45,7 +45,7 @@ struct snap_virtio_blk_queue_attr {
 	uint64_t			modifiable_fields;//mask of snap_virtio_blk_queue_modify
 	struct ibv_qp			*qp;
 	uint16_t			hw_available_index;
-	uint8_t				hw_used_index;
+	uint16_t			hw_used_index;
 
 	struct snap_virtio_queue_attr   vattr;
 };
